Move SetParameterProperties back to Database
